Validate forget password

See merge request xWildCAt/kdrmii!23

+++ app/controllers/userController.php
@@ -11,9 +11,9 @@ use core\View;
 class userController extends Controller
 {
     public function index(){}
-    /***
-     * *
-     * Password Reset route
+
+    /**
+     * Password reset route
      */
     public function passwordReset()
     {
@@ -48,13 +48,9 @@ class userController extends Controller
         }
     }
 
-    /***
-     * * User login
-     * Checks if fields fill out
-     * Checks if user information is correct
-     * Login
+    /**
+     * User login route
      */
-
     public function login(){
         if(User::isLogged()){
             indexController::redirect('/dashboard');
@@ -128,7 +124,7 @@ class userController extends Controller
         }
         if(intval($validation->valid_till) < time()){
             $user = new User($validation->user_id);
-            if($user->id !== null)
+            if($user->id !== null && $validation->type === "validate")
                 $user->delete();
             $validation->delete();
             indexController::moveToIndex("<div class='error'>Error: Validation not found.</div>");
@@ -140,10 +136,15 @@ class userController extends Controller
             $validation->delete();
             return;
         }
-        $user->validated = 1;
-        $user->save();
-        $validation->delete();
-        mailController::sendPassword($user->id);
-        indexController::moveToIndex("<div class='success'>User validated successful. Password is sent to your email.</div>");
+        if($validation->type === "validate"){
+            $user->validated = 1;
+            $user->save();
+            $validation->delete();
+            mailController::sendPassword($user->id);
+            indexController::moveToIndex("<div class='success'>User validated successful. Password is sent to your email.</div>");
+        }else{
+            mailController::sendNewPassword($user->id);
+            indexController::moveToIndex("<div class='success'>Your new password is sent to your email.</div>");
+        }
     }
 }
